# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AAELink HTTP API",
    "version": "0.0.41-alpha",
    "summary": "Auto-generated route inventory. Method/path surface only — request/response schemas not yet introspected."
  },
  "servers": [
    {
      "url": "https://example.com",
      "description": "Replace with your AAELink deployment URL"
    }
  ],
...
  "paths": {
...
    "/api/notifications/push": {
      "get": {
        "summary": "GET /api/notifications/push",
        "tags": [
          "notifications"
        ],
        "description": "_marker:_ external integration entry point (webhook / IdP / push provider / device)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "POST /api/notifications/push",
        "tags": [
          "notifications"
        ],
        "description": "_marker:_ external integration entry point (webhook / IdP / push provider / device)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "PUT /api/notifications/push",
        "tags": [
          "notifications"
        ],
        "description": "_marker:_ external integration entry point (webhook / IdP / push provider / device)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/search/files": {
      "get": {
        "summary": "GET /api/search/files",
        "tags": [
          "search"
        ],
        "description": "_marker:_ slack-compat surface (intentionally addressable, may be invoked by Slack-shaped clients)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "POST /api/search/files",
        "tags": [
          "search"
        ],
        "description": "_marker:_ slack-compat surface (intentionally addressable, may be invoked by Slack-shaped clients)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/team/preferences": {
      "get": {
        "summary": "GET /api/team/preferences",
        "tags": [
          "team"
        ],
        "description": "_marker:_ enterprise admin surface kept for parity (intentional, not yet wired into UI)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "POST /api/team/preferences",
        "tags": [
          "team"
        ],
        "description": "_marker:_ enterprise admin surface kept for parity (intentional, not yet wired into UI)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}